# Sage_agent - UIUX/Sage API Documents/swagger.products_services.json
{
  "info": {
    "title": "Sage Accounting API - Products & Services",
    "description": "Documentation of the Sage Business Cloud Accounting API.",
    "contact": {
      "name": "Sage Business Cloud Accounting API Team",
      "url": "mailto:sbc-accounting-api-developers@sage.com"
    },
    "version": "3.1"
  },
  "swagger": "2.0",
  "produces": [
    "application/json"
  ],
  "host": "api.accounting.sage.com",
  "basePath": "/v3.1",
  "schemes": [
    "https"
  ],
...
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Returns all Products",
        "description": "### Endpoint Availability\n\n* Accounting Plus: 🇨🇦, 🇬🇧, 🇮🇪\n* Accounting Standard: 🇬🇧, 🇮🇪\n* Accounting Start: 🇨🇦, 🇬🇧, 🇮🇪\n\n### Access Control Restrictions\n\nRequires the authenticated user to have any mentioned role in one of the listed areas:\n* Area: `Products & Services`: Read Only, Restricted Access, Full Access\n* Area: `Sales`: Read Only, Restricted Access, Full Access\n* Area: `Purchases`: Read Only, Restricted Access, Full Access",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "description": "Use this to filter by the item code or description.",
            "type": "string",
            "required": false
          },
          {
            "in": "query",
            "name": "updated_or_created_since",
            "description": "Use this to limit the response to Products changed since a given date (format: YYYY-MM-DDT(+|-)hh:mm) or date-time (format: YYYY-MM-DDThh:mm:ss(+|-)hh:mm). Inclusive of the passed timestamp.",
            "type": "string",
            "format": "date-time",
            "required": false,
            "example": "2015-10-27T16:24:52Z"
          },
          {
            "in": "query",
            "name": "deleted_since",
            "description": "Use this to limit the response to Products deleted since a given date (format: YYYY-MM-DDT(+|-)hh:mm) or date-time (format: YYYY-MM-DDThh:mm:ss(+|-)hh:mm). Not inclusive of the passed timestamp.",
            "type": "string",
            "format": "date-time",
            "required": false,
            "example": "2015-10-27T16:24:52Z"
          },
          {
            "in": "query",
            "name": "active",
            "description": "Use this to only return active or inactive items",
            "type": "boolean",
            "required": false
          },
          {
            "in": "query",
            "name": "items_per_page",
            "description": "Returns the given number of Products per request.",
            "type": "integer",
            "format": "int32",
            "default": 20,
            "minimum": 1,
            "maximum": 200,
            "required": false,
            "example": 20
          },
          {
            "in": "query",
            "name": "page",
            "description": "Go to specific page of Products",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "required": false
          },
          {
            "in": "query",
            "name": "attributes",
            "description": "Specify the attributes that you want to expose for the Products (expose all attributes with 'all'). These are in addition to the base attributes (name, path)",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all Products",
            "examples": {
             "application/json":
			 {
				"$total": 108,
				"$page": 1,
				"$next": "/products?page=2&items_per_page=10",
				"$back": null,
				"$itemsPerPage": 10,
				"$items": [
					{
						"id": "cf7e4478b34748ef95756e93a79e4059",
						"displayed_as": "Whiteboard - Drywipe (900 x 1200)",
						"$path": "/products/cf7e4478b34748ef95756e93a79e4059"
					},
					{
						"id": "22e9e0f4b53c4512ace0af975795b46a",
						"displayed_as": "Whiteboard - Drywipe (1000 x 1500)",
						"$path": "/products/22e9e0f4b53c4512ace0af975795b46a"
					},
					{
						"id": "9d1a5ae4c0d844f6a9d8336ddb2f3432",
						"displayed_as": "A4 Ledger Book - 5 Column",
						"$path": "/products/9d1a5ae4c0d844f6a9d8336ddb2f3432"
					},
					{
						"id": "a4aacfc4ef3f46a2a6193b50f0fb6b07",
						"displayed_as": "A4 Carbon Copy Book - Triplicate",
						"$path": "/products/a4aacfc4ef3f46a2a6193b50f0fb6b07"
					},
					{
						"id": "b86f251ec16e4697a8c22cb40d0d9f69",
						"displayed_as": "Shorthand Notebook - 80 Sheets",
						"$path": "/products/b86f251ec16e4697a8c22cb40d0d9f69"
					},
					{
						"id": "b5b7ace2b80342aa9fcc26c90d8421ce",
						"displayed_as": "Calculator - Desktop",
						"$path": "/products/b5b7ace2b80342aa9fcc26c90d8421ce"
					},
					{
						"id": "ccfb1cfc728e4f07839ce356bfca2ef2",
						"displayed_as": "Calculator - Pocket",
						"$path": "/products/ccfb1cfc728e4f07839ce356bfca2ef2"
					},
					{
						"id": "aa56dfd1c4f144e38ddd450cb8b98e9a",
						"displayed_as": "Calculator - Desktop (Printing)",
						"$path": "/products/aa56dfd1c4f144e38ddd450cb8b98e9a"
					},
					{
						"id": "04309417033d4bcc96e67ce21a304217",
						"displayed_as": "Calculator - Printing Rolls",
						"$path": "/products/04309417033d4bcc96e67ce21a304217"
					},
					{
						"id": "3fff912ee06c44808b7395eb74f8a24e",
						"displayed_as": "AT Mini Tower Case",
						"$path": "/products/3fff912ee06c44808b7395eb74f8a24e"
					}
				]
			}
            }
          }
        },
        "tags": [
          "Products"
        ],
        "operationId": "getProducts"
      },
      "x-sage-pathtitle": "Product collection",
      "x-sage-changelog": [

      ],
      "post": {
        "summary": "Creates a Product",
        "description": "### Endpoint Availability\n\n* Accounting Plus: 🇨🇦, 🇬🇧, 🇮🇪\n* Accounting Standard: 🇬🇧, 🇮🇪\n\n### Access Control Restrictions\n\nRequires the authenticated user to have any mentioned role in one of the listed areas:\n* Area: `Products & Services`: Restricted Access, Full Access",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Products",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/postProducts"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Creates a Product",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        },
        "tags": [
          "Products"
        ],
        "operationId": "postProducts"
      }
    },
...
  },
  "definitions": {
    "Base": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "The unique identifier for the item"
        },
        "displayed_as": {
          "type": "string",
          "description": "The name of the resource"
        },
        "$path": {
          "type": "string",
          "description": "The API path for the resource"
        }
      },
      "description": "### Endpoint Availability\n\n* Accounting Plus: 🇨🇦, 🇬🇧, 🇮🇪\n* Accounting Standard: 🇬🇧, 🇮🇪\n* Accounting Start: 🇨🇦, 🇬🇧, 🇮🇪",
      "x-sage-menusection": "Products & Services"
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "The unique identifier for the item"
        },
        "displayed_as": {
          "type": "string",
          "description": "The name of the resource"
        },
        "$path": {
          "type": "string",
          "description": "The API path for the resource"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "description": "The datetime when the item was created"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "description": "The datetime when the item was last updated"
        },
        "deleted_at": {
          "type": "string",
          "format": "date-time",
          "description": "The datetime when the item was deleted"
        },
        "deletable": {
          "type": "boolean",
          "description": "Indicates whether the product can be deleted"
        },
        "deactivatable": {
          "type": "boolean",
          "description": "Indicates whether the product can be deactivated"
        },
        "used_on_recurring_invoice": {
          "type": "boolean",
          "description": "Indicates whether the product has been used on a recurring invoice"
        },
        "item_code": {
          "type": "string",
          "maxLength": 255,
          "description": "The item code for the product"
        },
        "description": {
          "type": "string",
          "maxLength": 255,
          "description": "The product description"
        },
        "notes": {
          "type": "string",
          "maxLength": 500,
          "description": "The notes for the product"
        },
        "sales_ledger_account": {
          "$ref": "#/definitions/Base",
          "description": "The sales ledger account for the product"
        },
        "sales_tax_rate": {
          "$ref": "#/definitions/Base",
          "description": "The sales tax rate for the product"
        },
        "purchase_ledger_account": {
          "$ref": "#/definitions/Base",
          "description": "The purchase ledger account for the product"
        },
        "usual_supplier": {
          "$ref": "#/definitions/Contact",
          "description": "The usual supplier for the product"
        },
        "purchase_tax_rate": {
          "$ref": "#/definitions/Base",
          "description": "The purchase tax rate for the product"
        },
        "cost_price": {
          "type": "number",
          "format": "double",
          "description": "The cost price of the product"
        },
        "sales_prices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SalesPrice"
          },
          "description": "The sales prices for the product"
        },
        "source_guid": {
          "type": "string",
          "maxLength": 255,
          "description": "Used when importing products from external sources"
        },
        "purchase_description": {
          "type": "string",
          "maxLength": 250,
          "description": "The product purchase description"
        },
        "active": {
          "type": "boolean",
          "description": "Indicates whether the product is active"
        },
        "catalog_item_type": {
          "$ref": "#/definitions/Base",
          "description": "Indicates what the product type is"
        }
      },
      "description": "### Endpoint Availability\n\n* Accounting Plus: 🇨🇦, 🇬🇧, 🇮🇪\n* Accounting Standard: 🇬🇧, 🇮🇪\n\n### Access Control Restrictions\n\nRequires the authenticated user to have any mentioned role in one of the listed areas:\n* Area: `Products & Services`: Restricted Access, Full Access",
      "x-sage-menusection": "Products & Services"
    },
...
    "SalesPrice": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "The unique identifier for the item"
        },
        "displayed_as": {
          "type": "string",
          "description": "The name of the resource"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "description": "The datetime when the item was created"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "description": "The datetime when the item was last updated"
        },
        "price_name": {
          "type": "string",
          "description": "The name of the product sales price"
        },
        "price": {
          "type": "number",
          "format": "double",
          "description": "The sales price amount"
        },
        "price_includes_tax": {
          "type": "boolean",
          "description": "Indicates whether the sales price already includes tax"
        },
        "product_sales_price_type": {
          "$ref": "#/definitions/Base",
          "description": "The product sales price type"
        }
      }
    },
    "postProducts": {
      "type": "object",
      "properties": {
        "product": {
          "type": "object",
          "properties": {
            "description": {
              "type": "string",
              "description": "The product description"
            },
            "sales_ledger_account_id": {
              "type": "string",
              "description": "The sales ledger account for the product"
            },
            "purchase_ledger_account_id": {
              "type": "string",
              "description": "The purchase ledger account for the product"
            },
            "item_code": {
              "type": "string",
              "description": "The item code for the product"
            },
            "notes": {
              "type": "string",
              "description": "The notes for the product"
            },
            "sales_tax_rate_id": {
              "type": "string",
              "description": "The ID of the Sales Tax Rate."
            },
            "usual_supplier_id": {
              "type": "string",
              "description": "The ID of the Usual Supplier."
            },
            "purchase_tax_rate_id": {
              "type": "string",
              "description": "The ID of the Purchase Tax Rate."
            },
            "cost_price": {
              "type": "number",
              "format": "double",
              "description": "The cost price of the product"
            },
            "source_guid": {
              "type": "string",
              "description": "Used when importing products from external sources"
            },
            "purchase_description": {
              "type": "string",
              "description": "The product purchase description"
            },
            "active": {
              "type": "boolean",
              "description": "Indicates whether the product is active"
            },
            "catalog_item_type_id": {
              "type": "string",
              "description": "The ID of the Catalog Item Type."
            },
            "sales_prices": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "price_name": {
                    "type": "string",
                    "description": "The name of the product sales price"
                  },
                  "price": {
                    "type": "number",
                    "format": "double",
                    "description": "The sales price amount"
                  },
                  "price_includes_tax": {
                    "type": "boolean",
                    "description": "Indicates whether the sales price already includes tax"
                  },
                  "product_sales_price_type_id": {
                    "type": "string",
                    "description": "The ID of the Product Sales Price Type."
                  }
                }
              }
            }
          },
          "required": [
            "description",
            "sales_ledger_account_id",
            "purchase_ledger_account_id"
          ]
        }
      },
      "required": [
        "product"
      ],
      "description": "Creates a Product"
    },
...
  }
}